reptile_study/requirements.txt

python-whois
BuiltWith
HTTPie

pymysql
beautifulsoup4
lxml
requests
pymongo
redis
robobrowser
selenium
pillow
pytesseract
tesseract-ocr

nltk
scrapy

aiohttp
cchardet
aiodns